Split aligner_methods
